requirements.txt
kafka-python
pandas-msgpack
cython
scikit-learn
tqdm